requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
aiohttp>=3.9
python-dotenv>=1.0
jinja2>=3.1
itsdangerous>=2.1
